+++ services/importer/lib/importer/runner.rb
@@ -165,14 +165,14 @@ module CartoDB
           log_trace:  report
         ))
       end
-      
+
       def import(source_file, downloader, job=nil, loader_object=nil)
         job     ||= Job.new({ logger: log, pg_options: pg_options })
         loader = loader_object || loader_for(source_file).new(job, source_file)
         if(loader.respond_to?(:set_importer_stats))
           loader.set_importer_stats(@importer_stats)
         end
-        loader.options = @loader_options
+        loader.options = @loader_options.merge(tracker: tracker)
 
         raise EmptyFileError if source_file.empty?
 
@@ -254,7 +254,7 @@ module CartoDB
       attr_reader :results, :log, :loader, :stats
 
       private
- 
+
       attr_reader :pg_options, :unpacker, :available_quota
       attr_writer :results, :tracker
 
@@ -287,4 +287,3 @@ module CartoDB
     end
   end
 end
-