+++ services/table-geocoder/lib/internal-geocoder/admin1_column_polygons.rb
@@ -10,8 +10,8 @@ module CartoDB
       def search_terms_query(page)
         %Q{
           SELECT DISTINCT
-            quote_nullable(trim(#{@internal_geocoder.column_name})) as region,
-            quote_nullable(trim(#{@internal_geocoder.country_column})) as country
+            trim(quote_nullable(#{@internal_geocoder.column_name})) as region,
+            trim(quote_nullable(#{@internal_geocoder.country_column})) as country
           FROM #{@internal_geocoder.qualified_table_name}
           WHERE cartodb_georef_status IS NULL
           LIMIT #{@internal_geocoder.batch_size} OFFSET #{page * @internal_geocoder.batch_size}
@@ -26,13 +26,13 @@ module CartoDB
 
       def copy_results_to_table_query
         %Q{
-          UPDATE #{@internal_geocoder.qualified_table_name} AS dest
+          UPDATE #{dest_table}
           SET the_geom = orig.the_geom, cartodb_georef_status = orig.cartodb_georef_status
           #{CartoDB::Importer2::QueryBatcher::QUERY_WHERE_PLACEHOLDER}
           FROM #{@internal_geocoder.temp_table_name} AS orig
-          WHERE trim(dest.#{@internal_geocoder.column_name}::text) = orig.geocode_string
-            AND trim(dest.#{@internal_geocoder.country_column}::text) = orig.country
-            AND dest.cartodb_georef_status IS NULL
+          WHERE trim(#{dest_table}.#{@internal_geocoder.column_name}::text) = orig.geocode_string
+            AND trim(#{dest_table}.#{@internal_geocoder.country_column}::text) = orig.country
+            AND #{dest_table}.cartodb_georef_status IS NULL
           #{CartoDB::Importer2::QueryBatcher::QUERY_LIMIT_SUBQUERY_PLACEHOLDER}
         }
       end