+++ services/table-geocoder/spec/internal-geocoder/query_generator_factory_spec.rb
@@ -55,13 +55,13 @@ describe CartoDB::InternalGeocoder::QueryGeneratorFactory do
 
       query = query_generator.search_terms_query(0)
 
-      query.squish.should == 'SELECT DISTINCT(quote_nullable(trim(city))) AS city FROM "public"."untitled_table" WHERE cartodb_georef_status IS NULL LIMIT 5000 OFFSET 0'
+      query.squish.should == 'SELECT DISTINCT(trim(quote_nullable("city"))) AS city FROM "public"."untitled_table" WHERE cartodb_georef_status IS NULL LIMIT 5000 OFFSET 0'
     end
   end
 
   describe '#copy_results_to_table_query' do
     it 'should generate a suitable query to update geocoded table with temp table' do
-      @internal_geocoder.stubs('qualified_table_name').once.returns(%Q{"public"."untitled_table"})
+      @internal_geocoder.stubs('qualified_table_name').returns(%Q{"public"."untitled_table"})
       @internal_geocoder.stubs('temp_table_name').once.returns('any_temp_table')
       @internal_geocoder.stubs('column_name').once.returns('any_column_name')
       query_generator = CartoDB::InternalGeocoder::QueryGeneratorFactory.get(@internal_geocoder, [:namedplace, :text, :point])
@@ -69,11 +69,11 @@ describe CartoDB::InternalGeocoder::QueryGeneratorFactory do
       query = query_generator.copy_results_to_table_query
 
       query.squish.should == %Q{
-        UPDATE "public"."untitled_table" AS dest
+        UPDATE "public"."untitled_table"
         SET the_geom = orig.the_geom, cartodb_georef_status = orig.cartodb_georef_status
         #{CartoDB::Importer2::QueryBatcher::QUERY_WHERE_PLACEHOLDER}
         FROM any_temp_table AS orig
-        WHERE trim(any_column_name::text) = orig.geocode_string AND dest.cartodb_georef_status IS NULL
+        WHERE trim("any_column_name"::text) = orig.geocode_string AND "public"."untitled_table".cartodb_georef_status IS NULL
         #{CartoDB::Importer2::QueryBatcher::QUERY_LIMIT_SUBQUERY_PLACEHOLDER}
       }.squish
     end
@@ -82,7 +82,7 @@ describe CartoDB::InternalGeocoder::QueryGeneratorFactory do
   describe 'CDB-4269' do
     it 'should generate a suitable query generator for [:admin1, :column, :polygon]' do
       @internal_geocoder.stubs('column_name').twice.returns('region_column_name')
-      @internal_geocoder.stubs('qualified_table_name').twice.returns('any_table_name')
+      @internal_geocoder.stubs('qualified_table_name').returns('any_table_name')
       @internal_geocoder.stubs('country_column').twice.returns('country_column_name')
       @internal_geocoder.stubs('batch_size').twice.returns(10)
       @internal_geocoder.stubs('temp_table_name').once.returns('any_temp_tablename')
@@ -91,8 +91,8 @@ describe CartoDB::InternalGeocoder::QueryGeneratorFactory do
 
       query_generator.search_terms_query(0).squish.should == %Q{
         SELECT DISTINCT
-          quote_nullable(trim(region_column_name)) as region,
-          quote_nullable(trim(country_column_name)) as country
+          trim(quote_nullable(region_column_name)) as region,
+          trim(quote_nullable(country_column_name)) as country
         FROM any_table_name
         WHERE cartodb_georef_status IS NULL
         LIMIT 10 OFFSET 0
@@ -109,13 +109,13 @@ describe CartoDB::InternalGeocoder::QueryGeneratorFactory do
       }.squish
 
       query_generator.copy_results_to_table_query.squish.should == %Q{
-        UPDATE any_table_name AS dest
+        UPDATE any_table_name
           SET the_geom = orig.the_geom, cartodb_georef_status = orig.cartodb_georef_status
           #{CartoDB::Importer2::QueryBatcher::QUERY_WHERE_PLACEHOLDER}
           FROM any_temp_tablename AS orig
-          WHERE trim(dest.region_column_name::text) = orig.geocode_string
-            AND trim(dest.country_column_name::text) = orig.country
-            AND dest.cartodb_georef_status IS NULL
+          WHERE trim(any_table_name.region_column_name::text) = orig.geocode_string
+            AND trim(any_table_name.country_column_name::text) = orig.country
+            AND any_table_name.cartodb_georef_status IS NULL
             #{CartoDB::Importer2::QueryBatcher::QUERY_LIMIT_SUBQUERY_PLACEHOLDER}
         }.squish
     end